fix: issues during the bootstrap stage
arch/boot:    add missing mode switching from x86_64 compatibility
              mode to full 64-bits (long mode)

kpt_remap:    rework the kernel high-mem remapping for x86_64.
              isolate the remap logic of x86_64 and i386

gdt:          when loading gdt, the descriptor must follow the
              i386's format (all field are required), as processor
              will perform all the checks upon loading

idt:          remove the ist thing, we should reserve it for the
              "ABORT" interrupts to use.

ctx switch:   fix issue of a register being left in stack

vm:           add sign extend on virtual address to enforce consistency
              adjust the _PTE_X and introduce _PTE_NX
              minor adjustment on vm mapping for easy setup

kprintf:      fix the 64-bit integer not being correctly displayed
              in hex.

acpi:         fix the vague use of generic pointer type as acpi
              require all pointer type must be int32

variadic:     wrap the va_list unpacking method into arch-specific

makefile:     invoke correct QEMU depend on build architecture

lunabuild:    disable the gcc's SSE optimization when LConfig does
              enable this.

elf64:        typo fix, should check for CLASS64 flag set, rather
              than CLASS32

pmm:          pplist mapping granularity fix.

reactor:      some refactorings, clarify things

+++ lunaix-os/libs/klibc/itoa.c
@@ -5,12 +5,12 @@
 char base_char[] = "0123456789abcdefghijklmnopqrstuvwxyz";
 
 static char*
-__uitoa_internal(unsigned int value, char* str, int base, unsigned int* size)
+__uitoa_internal(unsigned long value, char* str, int base, unsigned int* size)
 {
     unsigned int ptr = 0;
     do {
         str[ptr] = base_char[value % base];
-        value = value / base;
+        value = value / (unsigned long)base;
         ptr++;
     } while (value);
 
@@ -27,11 +27,11 @@ __uitoa_internal(unsigned int value, char* str, int base, unsigned int* size)
 }
 
 static char*
-__itoa_internal(int value, char* str, int base, unsigned int* size)
+__itoa_internal(long value, char* str, int base, unsigned int* size)
 {
     if (value < 0 && base == 10) {
         str[0] = '-';
-        unsigned int _v = (unsigned int)(-value);
+        unsigned long _v = (unsigned long)(-value);
         __uitoa_internal(_v, str + 1, base, size);
     } else {
         __uitoa_internal(value, str, base, size);
@@ -41,7 +41,7 @@ __itoa_internal(int value, char* str, int base, unsigned int* size)
 }
 
 char*
-itoa(int value, char* str, int base)
+itoa(long value, char* str, int base)
 {
     return __itoa_internal(value, str, base, NULL);
 }
